修改&为and

+++ src/Application.py
@@ -34,6 +34,7 @@ try:
             continue
         ATListFileNames.append(f'{fileName}.txt')
 
+    enable_trace = 'n'
     if system_cate == 'Linux':
         enable_trace = input('是否抓取trace？（y/n）')
         if enable_trace == 'y':
@@ -73,18 +74,18 @@ def start_trace():
 
 try:
     print(f'Application process {os.getpid()}')
-    if system_cate == 'Linux' & enable_trace == 'y':
+    if system_cate == 'Linux' and enable_trace == 'y':
         multiprocessing.Process(target=start_trace).start()
     Application(port, baud_rate, ATListFileNames).run()
 except KeyboardInterrupt as ke:
     print()
     print("Exit...")
-    if system_cate == 'Linux' & enable_trace == 'y':
+    if system_cate == 'Linux' and enable_trace == 'y':
         os.kill(diag_pid, signal.SIGKILL)
     sys.exit()
 except serial.serialutil.SerialException as se:
     print(se)
-    if system_cate == 'Linux' & enable_trace == 'y':
+    if system_cate == 'Linux' and enable_trace == 'y':
         os.kill(diag_pid, signal.SIGKILL)
     if 'No such file or directory' in traceback.format_exc():
         print('输入的端口不存在')
